+++ src/main/java/lotto/model/domain/PublishedLotto.java
@@ -1,7 +1,6 @@
 package lotto.model.domain;
 
 import java.util.List;
-import lotto.Lotto;
 
 public class PublishedLotto {
