+++ src/main/java/lotto/model/service/LottoComparator.java
@@ -2,7 +2,7 @@ package lotto.model.service;
 
 import java.util.List;
 import lotto.model.domain.CompareResult;
-import lotto.Lotto;
+import lotto.model.domain.Lotto;
 import lotto.model.domain.PublishedLotto;
 import lotto.model.domain.WinningLotto;
