+++ src/main/java/lotto/view/OutputView.java
@@ -1,7 +1,7 @@
 package lotto.view;
 
 import java.util.Map;
-import lotto.Lotto;
+import lotto.model.domain.Lotto;
 import lotto.model.dto.ResultForView;
 import lotto.model.dto.PublishedLottoDto;
