+++ src/test/java/lotto/model/domain/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
